+++ mRemoteV1/App/Info/GeneralAppInfo.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
@@ -11,10 +11,10 @@ namespace mRemoteNG.App.Info
 {
     public static class GeneralAppInfo
     {
-        public const string UrlHome = "https://www.mremoteng.org/";
-        public const string UrlDonate = "https://mremoteng.org/contribute/";
-        public const string UrlForum = "https://www.reddit.com/r/mRemoteNG/";
-        public const string UrlBugs = "https://bugs.mremoteng.org/";
+        public const string UrlHome = "https://www.mremoteng.org";
+        public const string UrlDonate = "https://mremoteng.org/contribute";
+        public const string UrlForum = "https://www.reddit.com/r/mRemoteNG";
+        public const string UrlBugs = "https://bugs.mremoteng.org";
         public static string ApplicationVersion = Application.ProductVersion;
         public static readonly string ProductName = Application.ProductName;